requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
app_utils==0.1a4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
datasets==2.20.0
decorator==5.1.1
dill==0.3.8
Django==4.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
evaluate==0.4.2
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
gunicorn==22.0.0
huggingface-hub==0.22.2
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jiwer==3.0.4
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
openapi-codec==1.3.2
packaging==24.0
pandas==2.2.2
platformdirs==4.2.2
pooch==1.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
redis==5.0.3
referencing==0.35.0
regex==2024.4.16
requests==2.32.3
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
semantic-version==2.10.0
setuptools-rust==1.9.0
simplejson==3.19.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.5.0
sympy==1.12
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
tqdm==4.66.4
transformers==4.40.1
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
